Fixed more of the scripting

+++ HVCCC_Visualisation/Assets/Createjunctions.cs
@@ -105,7 +105,7 @@ public class Createjunctions : MonoBehaviour
             double jn_pos_x = topleft.transform.position.x + topleft_bottomright_length_x * (jn_lat_distance / lat_distance);
             double jn_lon_distance = junctions[i].zCoordinate - topleft_lon;
             double jn_pos_z = topleft.transform.position.z + topleft_bottomright_length_z * (jn_lon_distance / lon_distance);
-            float posy = Terrain.activeTerrain.SampleHeight(new Vector3((int)jn_pos_x, 0, (int)jn_pos_z));
+            float posy = Terrain.activeTerrain.SampleHeight(new Vector3((float)jn_pos_x, 0, (float)jn_pos_z));
             //Creates the junctions
             GameObject junction_object = Instantiate(junction, new Vector3((float)jn_pos_x, posy, (float)jn_pos_z), transform.rotation);
             junction_object.transform.parent = junction_parent.transform;
